python-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

# LLM
langchain-groq==0.1.9
langchain-core==0.2.38

# NLP
nltk==3.9.1
spacy==3.7.6
textblob==0.18.0